schemarepo: Schema repositories get their own package
It's a cross-cutting concern to need the dynamic value schemas of the
various plugins used with a Terraform configuration, but previously we
had the type for that in "package terraform", forcing lots of packages
to depend on the modules runtime even though they have no direct interest
in evaluating Terraform modules.

We'll now split it out into its own package "schemarepo". There are
currently forwarding aliases left behind in "package terraform" so that
we don't need to update all of the callers at once. We'll update the
callers whose dependency changes will have the most significant benefits
in future commits.

The new package is further split into a separate "laodschemas" part that
deals with the plugin interactions, because most packages that use a schema
repository expect that it was already constructed by someone else, and so
don't need or want to depend indirectly on all of the plugin execution
machinery.
