+++ internal/schemarepo/doc.go
@@ -1,0 +1,7 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: BUSL-1.1
+
+// Package schemarepo deals with the cross-cutting concern of gathering together
+// all of the schemas needed for manipulating data produced by the various
+// kinds of plugins.
+package schemarepo
