+++ internal/schemarepo/loadschemas/doc.go
@@ -1,0 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: BUSL-1.1
+
+// Package loadschemas knows how to load schemas from plugins to produce a
+// schema repository.
+package loadschemas
